+++ src/controllers/Repositories/AnimeThemesController.ts
